Fixed Sonar Cloud issues
