+++ app/src/main/java/com/github/se/studybuddies/database/DbRepository.kt
@@ -9,6 +9,7 @@ import com.github.se.studybuddies.data.DailyPlanner
 import com.github.se.studybuddies.data.Group
 import com.github.se.studybuddies.data.GroupList
 import com.github.se.studybuddies.data.Message
+import com.github.se.studybuddies.data.MessageVal
 import com.github.se.studybuddies.data.Topic
 import com.github.se.studybuddies.data.TopicFile
 import com.github.se.studybuddies.data.TopicFolder
@@ -87,7 +88,46 @@ interface DbRepository {
   suspend fun deleteGroup(groupUID: String)
 
   // using the Realtime Database for messages
-  fun sendMessage(chatUID: String, message: Message, chatType: ChatType, additionalUID: String = "")
+  fun sendMessage(chatUID: String, message: Message, chatType: ChatType, additionalUID: String) {
+    val messagePath = getMessagePath(chatUID, chatType, additionalUID) + "/${message.uid}"
+
+    val messageData =
+        mutableMapOf(
+            MessageVal.SENDER_UID to message.sender.uid, MessageVal.TIMESTAMP to message.timestamp)
+    when (message) {
+      is Message.TextMessage -> {
+        messageData[MessageVal.TEXT] = message.text
+        messageData[MessageVal.TYPE] = MessageVal.TEXT
+        saveMessage(messagePath, messageData)
+      }
+      is Message.PhotoMessage -> {
+
+        uploadChatImage(message.uid, chatUID, message.photoUri) { uri ->
+          if (uri != null) {
+            Log.d("MyPrint", "Successfully uploaded photo with uri: $uri")
+            messageData[MessageVal.PHOTO] = uri.toString()
+            messageData[MessageVal.TYPE] = MessageVal.PHOTO
+            saveMessage(messagePath, messageData)
+          } else {
+            Log.d("MyPrint", "Failed to upload photo")
+          }
+        }
+      }
+      is Message.FileMessage -> {
+        messageData[MessageVal.PHOTO] = message.fileUri.toString()
+        messageData[MessageVal.TYPE] = MessageVal.FILE
+        saveMessage(messagePath, messageData)
+      }
+      is Message.LinkMessage -> {
+        messageData[MessageVal.LINK] = message.linkUri.toString()
+        messageData[MessageVal.TYPE] = MessageVal.LINK
+        saveMessage(messagePath, messageData)
+      }
+      else -> {
+        Log.d("MyPrint", "Message type not recognized")
+      }
+    }
+  }
 
   fun saveMessage(path: String, data: Map<String, Any>)
